Working on a repository sub resource (problems with authentication filter however)
Changes to be committed:
	modified:   src/main/java/nl/tudelft/ewi/devhub/server/web/filters/RequireAuthenticatedUser.java
	modified:   src/main/java/nl/tudelft/ewi/devhub/server/web/resources/ProjectsResource.java
	new file:   src/main/java/nl/tudelft/ewi/devhub/server/web/resources/RepositoryResource.java
