+++ src/main/java/nl/tudelft/ewi/devhub/server/web/filters/RequireAuthenticatedUser.java
@@ -5,8 +5,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import javax.ws.rs.NameBinding;
+
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE, ElementType.METHOD })
+@NameBinding
 public @interface RequireAuthenticatedUser {
 
 }
